Change Duration to float 64

+++ crate.go
@@ -41,7 +41,7 @@ type endpointResponse struct {
 		Code    int
 	} `json:"error"`
 	Cols        []string        `json:"cols"`
-	Duration    int             `json:"duration"`
+	Duration    float64         `json:"duration"`
 	ColumnTypes int             `json:col_types`
 	Rowcount    int64           `json:"rowcount"`
 	Rows        [][]interface{} `json:"rows"`
